requirements.txt
# Web framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
websockets>=11.0.3

# LangChain & LLM libraries
langchain>=0.0.335
langchain-core>=0.1.1
langchain-community>=0.0.19
langchain-anthropic>=0.0.8
langgraph>=0.0.20
langgraph-checkpoint-sqlite>=0.0.2

# Tools
tavily-python>=0.2.2
openai>=1.3.8

# Testing
requests>=2.31.0
sseclient-py>=1.8.0
